+++ app.js
@@ -2,6 +2,6 @@ const { ScreenView } = require("./src/screen/screen_view");
 
 const screen = new ScreenView();
 while (screen.isRunning()) {
-  screen.update();
+  screen.render();
 }
 screen.unload();