+++ src/helpers/enums.js
@@ -1,4 +1,8 @@
 const directionEnum = { right: "right", left: "left", up: "up", down: "down", idle: "idle" };
+const gameStateEnum = { pause: "pause", restart: "restart", loading: "loading", update: "update" };
+const objectTypeEnum = { rect: " rectngle", circle: "circle" };
 module.exports = {
   directionEnum,
+  gameStateEnum,
+  objectTypeEnum,
 };