+++ src/models/collision_model.js
@@ -1,0 +1,4 @@
+// Make a model class for colisions that accepts a colision type enum
+// - Make a collision type function that returns true base on 2 rectangle
+// - Make a collision type function that returns true base on 2 Circle
+// - Make a collision type function that returns true base on 1 rectangle and 1 circle