+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-# Ninjutsu Clash 2022
+# Scrap to Hero 2022
 
 Updates of the game are listed here
 
 ## UPDATES
 
-### April 10 2022
+### April 10, 2022
 
 - Started the Ninjutsu Clash Open Source Project
+
+### April 13, 2022
+
+- Ninjutsu Clash was change to Scrap to Hero game